ArticleWordCloudAPI/update_articles/requirements.txt
newsapi-python
wordcloud